+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Report registration failure in ``ProcessTestCase``
 - Add a pseudo Python process to serve as a template
 - Add ``validate_urls`` method to ``BaseConnector`` class.
+- Validate connector settings on registraton
 
 Fixed
 -----